More Informative Error Message
Currently, when a bid is pending, you can not increase another bid unless you have the total amount in your balance. You have to wait for any pending claims to resolve before you can do this.

The current error message, in this case, can give the impression that the app is broken.